A sans-serif font family for body font
This was a bit of a difficult decision. One one hand, I perferred for the user
to select his/her own font via their browser settings. That said, the user still
has the option to override the font and use his/her own defaults.

This font choice is nearly identical to the current site font choice.