"Election"
The past few days of the DNC have demanded pause.  I am an Independent.  I
do not like Hillary Clinton. I am a Bernie supporter, and I was upset by his
endorsement of Hillary.  I had vowed not to vote for Hillary; I would
instead vote for Jill Stein.  The DNC, while very well done with a deeply
compelling facade, has not changed my perspective on Clinton.

It is perhaps said best by Bernie himself: "It's easy to boo, but it's
harder to look your kids in the face who would be living under a Donald
Trump presidency".  The conflict here is between my deep ideologies and
reality.  It's often said that a vote for Hillary is a vote against Trump;
such a perspective would shallow and purposeless.  But this isn't an
election for president---this is the most threatening assault on everything
I stand for that I hope I will ever witness in my lifetime.  To stand for
ideological purity would be to stand atop a mountain while the world around
me burns.  This is why Bernie chose to unite.

Should Trump win, my ideals that seem within reach could be blown back
decades.  As a matter of strategy, I cannot justify _not_ swallowing every
ounce of my pride.  Hillary's presidency is an unfortunate but necessary
consequence of the only permissible outcome.  I am not electing a president
of the United States.  I am electing _a United States_.

So I am doing what I never thought I would do: proposing that others too
factor this obscene equation and recognize how the very few remaining
variables affect the result.  My ideals continue to exist in part and in
spirit with Hillary as president.  With Trump, they are all but
vanquished.  Donald Trump must not be elected president of the United
States.  When (and if) you vote, think of it as a shot fired, not as a vote
cast.

"Election".

More information about my opinions on this topic can be found
[here][social-1] and [here][social-2].

[social-1]: https://social.mikegerwitz.com/conversation/21864
[social-2]: https://social.mikegerwitz.com/conversation/22026